tennisflow/src/pipeline.py
"""
TennisFlow Pipeline Module

This module orchestrates the full tennis swing analysis workflow, including:
1. Video frame extraction
2. Pose estimation with MoveNet
3. Temporal smoothing of keypoint sequences
4. Shot classification with RNN
5. Kinematic analysis and metrics calculation
6. Report generation
"""

import os
import cv2
import logging
import yaml
import numpy as np
from typing import Dict, List, Optional, Tuple, Union, Any

# Local imports
from .pose_estimation.movenet import MoveNetPoseEstimator
from .pose_estimation.smoothing import smooth_keypoint_sequence
from .classification.rnn_classifier import RNNClassifier
from .kinematics.segmentation import segment_swings
from .kinematics.analysis import calculate_metrics
from .utils.visualization import visualize_results

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class TennisAnalysisPipeline:
    """
    Main pipeline for tennis swing analysis.
    
    This pipeline processes a tennis video to:
    1. Extract pose keypoints from frames using MoveNet
    2. Apply temporal smoothing to the keypoint sequences
    3. Use an RNN to classify shot types
    4. Perform kinematic analysis to calculate biomechanical metrics
    5. Generate a comprehensive report
    """
    
    def __init__(self, config_path: str = "config.yaml"):
        """
        Initialize the pipeline with a configuration.
        
        Args:
            config_path: Path to the configuration YAML file
        """
        # Load configuration
        self.config = self._load_config(config_path)
        logger.info("Initialized Tennis Analysis Pipeline")
        
        # Initialize components (these will be properly implemented in subsequent steps)
        self.pose_estimator = None
        self.classifier = None
        
    def _load_config(self, config_path: str) -> Dict:
        """
        Load configuration from YAML file.
        
        Args:
            config_path: Path to the configuration file
            
        Returns:
            Configuration as a dictionary
        """
        try:
            with open(config_path, 'r') as f:
                config = yaml.safe_load(f)
            logger.info(f"Loaded configuration from {config_path}")
            return config
        except Exception as e:
            logger.error(f"Failed to load configuration: {e}")
            # Return default configuration
            return {
                "pose_estimation": {"model_type": "movenet_thunder"},
                "smoothing": {"method": "savgol", "savgol": {"window_size": 15, "polynomial_order": 3}},
                "pipeline": {"sample_rate": 2}
            }
            
    def initialize(self):
        """
        Initialize all components of the pipeline.
        """
        # Initialize pose estimator
        pose_config = self.config.get("pose_estimation", {})
        self.pose_estimator = MoveNetPoseEstimator(
            model_type=pose_config.get("model_type", "movenet_thunder"),
            confidence_threshold=pose_config.get("min_detection_confidence", 0.3)
        )
        
        # Initialize classifier if model path is provided
        rnn_config = self.config.get("rnn_classifier", {})
        model_path = os.path.join(
            self.config.get("paths", {}).get("models_dir", "models"),
            "rnn_classifier.h5"
        )
        if os.path.exists(model_path):
            self.classifier = RNNClassifier(
                model_path=model_path,
                sequence_length=rnn_config.get("sequence_length", 30),
                classes=rnn_config.get("classes", ["forehand", "backhand", "serve", "neutral"])
            )
        else:
            logger.warning(f"RNN classifier model not found at {model_path}. Classification will be disabled.")
        
        logger.info("Pipeline components initialized")
        
    def process_video(self, video_path: str) -> Dict[str, Any]:
        """
        Process a tennis video through the entire pipeline.
        
        Args:
            video_path: Path to the video file
            
        Returns:
            Analysis results including detected swings and metrics
        """
        logger.info(f"Processing video: {video_path}")
        
        # Initialize components if not already done
        if self.pose_estimator is None:
            self.initialize()
        
        # Extract frames and keypoints
        keypoint_sequences, frame_timestamps = self._extract_keypoint_sequences(video_path)
        
        if not keypoint_sequences:
            logger.error("Failed to extract keypoint sequences from video")
            return {"error": "No keypoints detected in video"}
        
        # Apply temporal smoothing
        smoothed_sequences = self._apply_smoothing(keypoint_sequences)
        
        # Detect and classify swings
        swings = self._detect_and_classify_swings(smoothed_sequences, frame_timestamps)
        
        # Calculate kinematic metrics for each swing
        results = self._calculate_metrics_for_swings(swings, smoothed_sequences, frame_timestamps)
        
        logger.info(f"Completed processing video: {video_path}")
        return results
    
    def _extract_keypoint_sequences(self, video_path: str) -> Tuple[List[np.ndarray], List[float]]:
        """
        Extract keypoint sequences from video frames.
        
        Args:
            video_path: Path to the video file
            
        Returns:
            Tuple of (keypoint_sequences, frame_timestamps)
        """
        # This is a placeholder implementation
        # The actual implementation will:
        # 1. Extract frames from video at specified sample rate
        # 2. Run MoveNet pose estimation on each frame
        # 3. Collect keypoints into sequences
        
        logger.info("Extracting keypoint sequences (placeholder)")
        return [], []
    
    def _apply_smoothing(self, keypoint_sequences: List[np.ndarray]) -> List[np.ndarray]:
        """
        Apply temporal smoothing to keypoint sequences.
        
        Args:
            keypoint_sequences: List of raw keypoint sequences
            
        Returns:
            List of smoothed keypoint sequences
        """
        # This is a placeholder implementation
        # The actual implementation will apply Savitzky-Golay or Kalman filtering
        
        logger.info("Applying temporal smoothing (placeholder)")
        return keypoint_sequences
    
    def _detect_and_classify_swings(
        self, 
        smoothed_sequences: List[np.ndarray],
        frame_timestamps: List[float]
    ) -> List[Dict]:
        """
        Detect and classify tennis swings.
        
        Args:
            smoothed_sequences: List of smoothed keypoint sequences
            frame_timestamps: Timestamps for each frame
            
        Returns:
            List of detected swings with classification
        """
        # This is a placeholder implementation
        # The actual implementation will:
        # 1. Segment the sequences into potential swings
        # 2. Classify each swing using the RNN
        
        logger.info("Detecting and classifying swings (placeholder)")
        return []
    
    def _calculate_metrics_for_swings(
        self,
        swings: List[Dict],
        smoothed_sequences: List[np.ndarray],
        frame_timestamps: List[float]
    ) -> Dict[str, Any]:
        """
        Calculate kinematic metrics for each swing.
        
        Args:
            swings: List of detected swings
            smoothed_sequences: List of smoothed keypoint sequences
            frame_timestamps: Timestamps for each frame
            
        Returns:
            Analysis results including swings and metrics
        """
        # This is a placeholder implementation
        # The actual implementation will calculate biomechanical metrics for each swing
        
        logger.info("Calculating metrics for swings (placeholder)")
        return {
            "swings": swings,
            "metrics_summary": {}
        } 