tennisflow/requirements.txt
# Core dependencies
numpy>=1.20.0
scipy>=1.7.0
pandas>=1.3.0
matplotlib>=3.4.0
pyyaml>=6.0

# Machine Learning
tensorflow>=2.8.0
tensorflow-hub>=0.12.0
scikit-learn>=1.0.0
keras>=2.8.0

# Computer Vision
opencv-python>=4.5.0
pillow>=8.0.0

# Data processing
tqdm>=4.62.0

# Web API
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.0

# Testing
pytest>=6.2.0

# Development tools
jupyter>=1.0.0
notebook>=6.4.0
ipykernel>=6.0.0 